[REFACTOR] AWS S3 에 접근 하는 로직을 Repository 단으로 모듈화
기존에는 Service에서 바로 aws s3로 접근하는 로직이 있었음.
하지만 해당 부분이 s3 대신 파일 시스템이 될 수도 있고 다른 저장소를 사용할 수 있기 때문에 추상화를 적용했고 Repository 계층을 만들고 구현체로 aws s3 저장소를 만들어 모듈화함.
