+++ src/main/java/com/ssg/usms/business/store/dto/CctvDto.java
@@ -1,4 +1,4 @@
-package com.ssg.usms.business.store;
+package com.ssg.usms.business.store.dto;
 
 import lombok.Getter;
 import lombok.Setter;
