+++ src/main/java/com/ssg/usms/business/store/service/CctvService.java
@@ -1,5 +1,6 @@
-package com.ssg.usms.business.store;
+package com.ssg.usms.business.store.service;
 
+import com.ssg.usms.business.store.dto.CctvDto;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
