+++ src/main/java/com/ssg/usms/business/video/repository/VideoRepository.java
@@ -1,0 +1,6 @@
+package com.ssg.usms.business.video.repository;
+
+public interface VideoRepository {
+
+    byte[] getVideo(String key);
+}
