+++ src/main/java/com/ssg/usms/business/video/service/StreamKeyService.java
@@ -1,8 +1,8 @@
 package com.ssg.usms.business.video.service;
 
 
-import com.ssg.usms.business.store.CctvDto;
-import com.ssg.usms.business.store.CctvService;
+import com.ssg.usms.business.store.dto.CctvDto;
+import com.ssg.usms.business.store.service.CctvService;
 import com.ssg.usms.business.video.dto.HttpRequestCheckingStreamDto;
 import com.ssg.usms.business.video.exception.AlreadyConnectedStreamKeyException;
 import com.ssg.usms.business.video.exception.ExpiredStreamKeyException;
