+++ src/test/java/com/ssg/usms/business/video/service/StreamKeyServiceCheckingAuthOfPushingStreamTest.java
@@ -1,7 +1,7 @@
 package com.ssg.usms.business.video.service;
 
-import com.ssg.usms.business.store.CctvDto;
-import com.ssg.usms.business.store.CctvService;
+import com.ssg.usms.business.store.dto.CctvDto;
+import com.ssg.usms.business.store.service.CctvService;
 import com.ssg.usms.business.video.dto.HttpRequestCheckingStreamDto;
 import com.ssg.usms.business.video.exception.AlreadyConnectedStreamKeyException;
 import com.ssg.usms.business.video.exception.ExpiredStreamKeyException;
